+++ NEWS.md
@@ -1,3 +1,15 @@
+# prodigenr 0.3.0
+
+## New features/functionality
+
+* `prodigen()` now creates a new project structure that mimics R package development
+more, specifically `devtools` (since it is well documented and actively
+maintained) (#18, #15, #14). The old style has been removed.
+* Added a function to include a [STROBE](http://www.strobe-statement.org/index.php?id=strobe-home) 
+checklist for  reporting guidelines (#21).
+* Added function to include an `rfigshare` script to send to [figshare](https://figshare.com/) (#19).
+* Added function to include a more complete MIT license (#20).
+
 # prodigenr 0.2.7
 
 * Added a `NEWS.md` file to track changes to the package.